requirements.txt
requests>=2.28.2
psycopg2-binary
requests_oauthlib==0.5.0
Werkzeug==0.16.0
celery==5.2.2
Flask==1.1.2
Jinja2==2.11.3
MarkupSafe==1.1.1
itsdangerous==1.1.0
python-dotenv==0.14.0
bcrypt
sendgrid==6.9.7
Flask-Markdown==0.3
gunicorn==20.1.0
redis==4.5.1
Flask-Compress==1.13
Pillow>=10.0.0
pytest>=7.0.0
sentry-sdk>=1.13.0
